Improve persist query params functionality
Rename WHITELISTED_QUERY_PARAMS configuration to
PERSIST_QUERY_PARAMS. This is more descriptive of what it does.

Replace the falsy/truthy check on the query object with undefined
checks. This will make sure the query param is also persisted for
falsy values, such as 0 or null (when you don't pass a value,
vue-router will use null)