Fix import order in bootstrap.js
Put the polyfill script in the entry configuration, to make sure
polyfills are loaded before all other js.

Put the settings file before all application code in the import
order. For now this seems to fix issues with bundling order, but
we might have to investigate if this will always guarantee the
correct order of bundling