Restructure webpack configuration
This commit is only supposed to restructure the way we write
webpack configuration. There should be no change in the actual
contents of the configuration.

The file structure has changed to make it clear where to
look when looking for a specific piece of config:
 - webpack.conf.dev.js and webpack.conf.prod.js now only serve
 as entry files. They do not contain configuration
 - webpack.conf.base.js only contains top-level configuration
 properties
 - all other configuration has been split into the top-level
 objects of a webpack config file to webpack.partial.conf.<prop>.js

The filenames have been changed to start with webpack.conf instead
of ending in .conf. This will make sure that all the
webpack.partial.conf files are next to each other when sorted
alphabetically.

All helper functions (mostly for loaders) have been moved to be
inline at the location where they are used. This will make sure
all the configuration related to a single property is next to
each other (as opposed to split among different modules)

A small number of conf options have been removed that are already
automatically set by webpack when using the 'mode' property.

Comment blocks have been added in several places to create a
visual separation in long configuration files.