Restructure config.js and introduce buildTypes
Instead of passing an isDevelopment flag to webpack.conf.base.js,
we now pass a buildType string which (by default) can be either
'development' or 'production'. This leaves room for more complex
build setups where more types of builds are desirable. The partial
files will now also receive the `buildType` string. Because the
buildType in most cases is either development or production,
the `isDevelopment` boolean is also passed to the partials for
convenience.

The `dev` and `build` objects of `config.js` have been slightly
refactored and renamed for clarity.
 - because `env` exists in both of these, it is moved one level
 up. The distinction between build types is made inside of the
 `env` object instead
 - `dev` has been renamed to `devServer`, to indicate that these
 options are specifically meant for the webpack development
 server (as opposed to development builds, which can theoratically
 also be deployed).
 - `build` has been renamed to `dist` (technically a dev server is
 also a "build", so that term is a bit too generic)