Disable strict-boolean-expressions
I notice that myself and other devs often run into this rule failing. It's not always obvious how to make the linter happy, and the fix sometimes looks somewhat convoluted. I am also not entirely convinced that this rule is very useful, but if you think otherwise please leave a comment. It seems that tslint-config-airbnb has also removed the rule due to there being some issues with it.