Add "skeleton" field to package.json
Create a "skeleton" field that mirrors the "name" and "version" properties.

The reasoning behind this: when you create your own project, you should be able to change the "version" and "name" field in `package.json` to reflect your project rather than `vue-skeleton`. If we do that however, we lose information about which skeleton version the project was originally started with.