Remove setting NODE_ENV in package.json scripts
This does not seem to be necessary. Webpack already sets `NODE_ENV` in our source code using the define plugin. As far as I know there isn't any code in the build script or `webpack-dev-server` that relies on `NODE_ENV` being set. That allows us to remove the `cross-env` dependency.